requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
replicate==0.15.5
gTTS==2.4.0
deep-translator==1.11.4
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2==2.9.9